Prefer Rails 6 ActiveRecord::DatabaseConfigurations API when available
Before Rails 6, `ActiveRecord::Base.configurations` was a hash
which `current_config` returned an entry from. In Rails 6 and later,
it's a `ActiveRecord::DatabaseConfigurations` object which has a
`configs_for` method which gives us the the configuration for an
environment.

The old hash API is still available in Rails 6.1 but is deprecated.